test(messaging): split Telegram diagnostic coverage
